add PhantomData to ensure correct marker traits
AtomicPtr doesn't propagate marker traits, e.g. Send and Sync.  This
allows soundness bugs because FreezeBox was not overriding the AtomicPtr
behavior.

The easiest way to fix this is to drop in a PhantomData member, which
will propagate marker traits.

Add compile-fail doctests that ensure non-Send and non-Sync types don't
magically become Send and Sync when wrapped in a FreezeBox.