change new() to accept Option<T>
It's desirable (e.g. for unit tests) to have some way of creating
a pre-initialized FreezeBox. I played with a few ideas, and eventually
landed on just changing new() to be new(val: Option<T>).

I think this matches most expectations about what new is for, and it
will be less confusing than having two or three different initialization
functions. I still expect default() to be the common case, so I'm not
that bothered by the extra typing in "FreezeBox::new(Some(123))".